app/requirements.txt
Flask==2.0.2
python-dotenv
requests==2.28.2
gunicorn==20.1.0
llama-index==0.4.26
openai==0.27.1
html2text==2020.1.16
feedparser==6.0.10
validators==0.20.0
langdetect==1.0.9
youtube-transcript-api==0.5.0
grpcio==1.53.0
grpcio-tools==1.53.0
protobuf==4.22.1
pydub
lxml